refactor: simplify Header and add auth controls
